Fixed checkAnswer to support doubles

+++ src/Question.java
@@ -23,7 +23,7 @@ public class Question {
     
     public boolean checkAnswer(JTextField textField){
         correct = false; 
-        int userNum = Integer.valueOf(textField.getText());
+        Double userNum = Double.valueOf(textField.getText());
         if (userNum == answer){
              correct = true; 
         }
